+++ PinpadLibrary/src/main/java/io/cloudwalk/pos/pinpadlibrary/internals/utilities/PinpadUtility.java
@@ -618,7 +618,7 @@ public class PinpadUtility {
         String RSP_ID = input.getString(ABECS.RSP_ID, "UNKNOWN");
 
         switch (RSP_ID) {
-            // case ABECS.OPN: response = OPN.buildResponseDataPacket(input); break;
+            case ABECS.OPN: response = OPN.buildResponseDataPacket(input); break;
             // case ABECS.CHP: response = CHP.buildResponseDataPacket(input); break;
             // case ABECS.GPN: response = GPN.buildResponseDataPacket(input); break;
 
@@ -731,8 +731,6 @@ public class PinpadUtility {
         int response = 0;
 
         if (input.length >= length) {
-            Log.h(TAG, input, length);
-
             for (int i = length - 1, j = 0; i >= 0; i--, j++) {
                 if (input[j] < 0x30 || input[j] > 0x39) {
                     String message = String.format(US, "getIntFromDigitsArray::input[%d] [%02X]", j, input[j]);
@@ -742,8 +740,6 @@ public class PinpadUtility {
 
                 response += (input[j] - 0x30) * ((i > 0) ? (Math.pow(10, i)) : 1);
             }
-
-            Log.d(TAG, String.format(US, "getIntFromDigitsArray::response [%d]", response));
         } else {
             String message = String.format(US, "getIntFromDigitsArray::input.length [%d] length [%d]", input.length, length);